Updated Macaroni and its friends to use Boost 1.46.1.  Of course a million things changed in the filesystem library since I was using depreciated methods, so I had to update some code in Macaroni (and in the boost generator).  Also I was getting really strange errors before I remembered that duh, you have to use Boost Build 1.46.1 (and not the prior version) if you want to compile against the 1.46.1 libraries.
Also added a fix so that generation of classes in the root namespace works okay, instead of writing out "namespace %ROOT%" in the generated source.  Putting stuff in the root namespace ain't a great idea normally but for tests and examples it makes it much easier to analyze the results.

Began to write a test for constructor overloading.  I need to implement operators next.