Made ClassParent use the LuaGlue generator.  Had to fix LuaGlue to not generate glue for classes with "LuaGlueClass" already defined in their Attributes.  Also added a lot to the Dependencies.mcpp to tell Macaroni about manually generated C++ code to annotate it with Lua info.  After all this work I was able to get out of writing one Lua glue class.  I think that's worth it!  The Lua Glue generator needs some work, of course.  It would be cool to be able to generate properties as well as len overloads.  But I'll need to sit down and think about that, and that was not my focus today.
Changed ClassParent to use "Type" instead of "Node" to refer to the parent, since really in C++ what you put is the type you inheret from (which is possibly defined on the spot) and not a specific class.  Added Lua Glue for ClassParentList.

Note- there are ENORMOUS memory leaks in some of these list structures.  It's going to be a tough time fixing them all.  Many times a list of smart ptrs is used by something which exists inside the Context.  Its pretty bad.

So, another pretty big thing- I've finally added inheritance!  I knew if I sat down, gritted my teeth and nearly gave myself a heart attack this morning I could do it.  I'm not sure if its 100% or not- I've only written an integration test for it, not any unit tests, and I'm certain I've missed some things- but thank God its finally in.  Putting this feature in, even in its current state, makes me feel like a load has been lifted off my mind.  I guess its only take over a year to add this one feature. I'm so happy I don't even care that it's probably broken in ways I will only discover later.

I guess the next big change is to make the Lua Glue generation nicer.

Btw, to compile Macaroni right now I've been in using the following command:

macaroni generatorPath ..\Generators build

This is because I have had to update the LuaGlue stuff to work on Macaroni as I was adding the inheritance feature.  But ironically the generators right now have stuff in them to write the C++ code which causes Macaroni not to work.  So in ClassHFileGenerator I have a variable named "FUTURE" which I'm setting to true or false depending on if I need to compile Macaroni or run the inheritance test.  So to get Macaroni to work, you need to point it to the current working copy of the Generators directory and also set FUTURE = false in ClassHFileGenerator.lua.