I updated the stand-alone Lua dependencies to correctly build as an exportable MS DLL if needed by adding "LibraryRequires" to the Boost Build generator.
Turned out that friend definitions were linking to the Function instead of the specific function overload.  This caused some major ambiguity problems because it meant the Node was friends with *all* of the function overloads.  For intrusive_ptr friends this was a big issue, because two classes would foster global methods, but then each one would be friends with all the other ones.  So I changed it so ~friend in a function def would make the class friends with just that overload, where the standard "friend [node]" definition would make it friends with every function overload, and also put a comment in the header file that this was probably not desired and explain how to do it differently.

Added better error messages for when something goes wrong in LuaEnvironment.

Changed AttributeTableLua to catch exceptions and raise them as Lua errors.

I- *almost* - very easily changed Macaroni to allow for a new kind of access called "hidden".  Basically, the CPP file includes the entire contents of the H file but its slightly different and may include extra bits.  You can think of it as more private than private, but the real treat is if you add methods to a files which are just private anyway you can declare them as "~hidden" and not have to recompile everything which includes the H file.  But to change this, I realized the enum for Attributes would need to evolve a bit.

So I changed Attributes in a few substantial ways.  First off, the Pippy Parser will now add them to functions and typedefs.  I also realized that currently I put the attributes after the definition, but in the future may want to change it to be more like Java and .NET and place them before it.  Because of this I removed the ability to put them in the midst of a class and made it possible to put them right after a class definition and before its braces.

I also made it possible to declare an attribute without a value.  Second, it turned out that allowing plain code blocks as attribute string values was ambigious because there was no way to know if they were for the function or class body following the attribute or part of the attribute itself, so I changed Attributes to force an equal sign before a class body following an attribute name.  I also had allowed a semicolon to follow attributes, but I removed this because it made working with typedefs impossible (unless you ended them with two semicolons, which would have been super lame).

Changed the Attribute test just a bit so the Lua code which gets called functions more as a test for the Pippy Parser and Macaroni::Model and throws error if the Attribute values aren't as expected (before it just printed them out to the benefit of no one).

Changed Access from an enum to a strange creature resembling a Java enum but with functions to grab what in other languages might be static values.  Discovered to me sadness that I would have to hand-write Lua glue for them, so instead decided to focus on the Lua Glue generator.

Made some substantial progress on the Lua glue stuff.  I added the ability to give a C++ method an attribute marking it as a Lua property.  So if you have a getter in C++, the Lua metatable's index method will return a value if a variable named after some name (which is defined by the attribute you use) is accessed.  I also added setters.  I also changed it so static methods go to the Lua global table now and can't be accessed from the instance, and they can also have properties too, but these can only be accessed, not set.  I made it so setting any value which isn't there throws an exception.  Finally, I wrapped all the C++ Lua Glue code in try / catches which raise Lua errors.  All of this is tested in Tests/Features/Lua.  That test allowed me to catch a lot of issues with the ~friend keyword which I mention above, and forced me to add and fix issues with Attributes.

Added std::exception to the Macaroni::CppStd::2003 library.

Also changed the BoostBuild test to include timings because I was curious if the new ability to build things as libraries was helping compile time.  The sad truth is it doesn't seem to be (in fact in some of the tests building dependencies as DLLs takes longer) but maybe if I increase the size of what I'm building I'll see better results.