This pertains to 0.1.0.15.  While compiling this I had to use 0.1.0.14 while using the current Lua generators. A namespace header wasn't being created correctly.  Scope.h wouldn't get parsed due to how the includes were set up, so Class.h would have an error saying Scope was not defined.  Because of this I had to do two things. 1. I added an attribute called Macaroni::Cpp::UseLightDef.  When the CPP generators see this, they know its OK to try to figure out the light definition of a Node even if that node defines an H file.  Before this if you gave a class an hfile the CPP generator wouldn't know how to represent it, even if it was as simple as "class blah;".
I fixed up the Access class a bit.  I now generate its glue with the Lua glue generator, because hey why not.  It actually is doing a really good job.

I changed Access to AccessPtr in a lot of places.  I very quickly began to see this as a mistake.  I'll probably change it back to just "Access" soon instead of the smart pointer.

Removed the hand-crafted ClassParent h and cpp files and now generate them with Macaroni following the ~hfile work I described above.

Changed a lot of the Glue code to use the new Access properties and methods to determine things like header and library visibility, but had to include guards for b/c with older Macaroni versions.

Added the ability to add operator methods to the metatables in the generated Lua glue.

Bumped up version to 0.1.0.15.

Wrote a bug recreate for the CPP generator error I described above.