Did some thinking and realized much of the problems I've been experiencing with Macaroni generated typedef headers being difficult to use is because of the dumb policy of putting all of them in the namespace underscore header. The more I thought about it the more difficult it seemed to solve the problem and there was no benefit to doing it that way at all, so I changed it. Now all type defs are treated the same way classes are- they get their own header file. This means there's far more headers than before but it let me figure out some of the generation issues and fix them one-by-one.
Made the Lua generator create two header files so the "RegisterInternalModules.cpp" file wouldn't have to be updated everytime a new Node was marked to have Lua Glue added.

Added a new type to the TypeModifiers: "Light." This is used to specify that a type's full definition is not needed even though it appears to be used (useful for shared_ptr stuff and other places). Made the TypeModifiers property of Type accessible from Lua glue. It's specified by putting '~l' after other type modifiers.

Added the Boost library staging path to the generated Boost Build script for Macaroni. This makes the release version compile in Windows.

Added a property "CPP_GENERATE_VERBOSE" to the CPP generator. Its in the code now, but this could be changed fairly easily. If set, it makes the generated code verbose to track down issues in the generator. By adding this I felt comfortable removing a lot of nasty cruft and the generated source looks a tiny bit better.

Discovered a bug thats unresolvable, so I created a new "bug" project for it.

Cleaned up a lot of code.