Adding "Platforms" property to targets.
This was necessary to distinguish Windows and Dreamcast targets.
Simply add "platform=blah" to a unit definition. BoostBuild2 uses
this by checking extra arguments. Currently it adds an additional
usage requirement to the created target, making it so the target
will only compile then (the idea is this matches the platform).

Also added the ability to specify "top" and "bottom" blocks for
plain unit files, though this turned out to be not that big of
a deal given the new feature I describe above.