Multiple fixes and advances.
Introduced "generateLess".

This allows only certain elements to be generated, such as C++ code,
while omitting other elements such as project files used for IDEs.


Limited profiling support.

Introduced a way to measure Macaroni's execution time. Using Window's
high-resolution timers, the time since the program began is printed out to
the console. The total running time of the program is given when the
program finishes.


Resurrected the InterfaceMh plugin.

This plugin creates a simple bit of Macaroni source code that omits all
important information about a class except the .h file that gets
generated for it. Reading these is thus quicker than re-parsing the
source.


"loadingAsDependency" feature

A new global variable was introduced to the project system which allows
the Lua scripts governing how a project is parsed to know if the
project is being loaded as a dependency or not. If it is, it is possible
to read the Interface file by itself rather than iterate all of the
Macaroni code again, and exe and test targets are ignored. This has to
be supported by the project though. Currently works in the SimpleProject
script.
This feature took the start up time of a project that utilized three
dependent projects (as in the time before the generation phase) down
from  1922ms to 473ms.


Fixed a nullptr bug in the Scope class.


Added some library files for Boost v.1.55 dependencies.