Accept build options from command line
This allows users to specify options to bjam from the command line.
It works by accepting a list of strings with the form "-o" or
"-buildOptions" which are then added to the global "properties"
variable as an array to the key "buildOptions".

This involved work to pass other stuff to the properties variable
which should make things easier in the future.

Also moves the Lua code to call Boost Build out of the SimpleProject
function and into the BoostBuild2 plugin, where it belongs.