Finally allow "struct" in addition to "class"
Basically a struct is a class in Macaroni with the "IsStruct" property
set to true.

In the process, I had to majorly change the Parser Lua glue which is
used in the tests to pass in a Target instead of a Library, which is
the deprecated way of parsing stuff.