Allow Functions to own their own Targets
This fixes a long standing problem where global functions could not
be in their own targets and thus always had to be "owned" by a class,
which meant units couldn't have free standing functions in many cases
unless they were adopted nodes that used the ~global keyword (though it
seemed to work sometimes... it was messed up to say the least).

This commit attempts to fix that and almost undoubtedly will introduce
some hellacious new bug.