More block related changes
So I left off and didn't commit and can't quite recall where I was. I
think the next step of this is to make sure that "top" and "bottom"
blocks of classes are merely adopted by them but the actual nodes where
the blocks are hosted is nested under the namespace not the class, and
the target is whatever unit target they need to be. So if its in the
class then its the class target.

The more I thought about it, it seems like the unit file generation
should also be in C++ as the logic is pretty critical to the goal of
Macaroni itself, and much of the block logic was designed for convience
of the Lua implementation. :(