This commit makes the generator check for Cycles
It turns out that in this new golden future, not putting a ~l type
modifier on, say, a typedef which only needs the class forward
declaration will consistently produce an error- which is pretty cool!
The new generator will print out a somewhat garish error message but
one that contains the exact thing that has to be fixed. Neat.