Correctly graph nested classes
This commit inverts a recent change. Now instead of a nested class
depending on its enclosing class, the enclosing class pulls in all
of the dependencies of the nested class. Since the enclosing class
is responsible for writing the nested class this seems to work, and
also makes it so the enclosing class depends on any type definitions
that the nested class needs.