Getting rid of ~global and ~blocks reliant tests
It is painful to think of how long I must have spent asserting logic in
Macaroni which in retrospect was all part of an enormous design mistake.
Many of the tests try to assert arbitrary like rules of how dependencies
worked in the presence of ~globals living in units.

This commit also fixes a number of issues which were bugs in Macaroni.
For example, in one case a class had a dependency on a class with the
same name, but did not use it. Previously no using statement had been
written, but now one was, which triggered an error message about how
the name was defined twice. Which makes me really wonder why the test
did that in the first place.

There's also a weird case where an example of the auto and decl keyword
were not causing a compiler error, which makes me wonder if that code
had yet passed (maybe I wrote it when I began working on Macaroni IV
before the tests had been run again).