Write out include file for typedef for fd
This feels so wrong but there was code in the tests that wasn't
compiling because a class rightfully depended only on a forward decl
for a typedef, but that can't be written, so the code wouldn't compile.
The logic in the CodeGraph class seemed sound, so the only choice
was to change the CodeTextProducer to output the include file directive
for the forward decl. Time will tell if this was the right choice.