Fixes MhFileGenerator
Now when a library is "installed" in Macaroni, the old Interface.mh
file is replaced with a file named lib.mh, where "lib" is the library
name. However instead of the old "class { ~hfile=..." bologna it uses
the new ~extern element type. It took some noodling but it looks to
be working correctly.

Next, there's a new language feature that allows "externs" to be
exported. Normally you wouldn't want this as ~externs just get around
issues where Macaroni code can't know what some type is, but sometimes
when authoring a library you might use an extern intentionally and
want to make it available elsewhere. So now adding "export" inside of
an "~extern" def after the node name will make the interface of the
library include it. Doing this also required having an extern know which
target created it- if "export" is not specified, the extern doesn't
believe itself to be owned by any target.

Finally, if an extern has a using statement with all whitespace it
is ignored by the PippyParser.