Fix generation of ~global items
Even though the ~global keyword is kind of embarassing, the ability to
define a global function which is a friend of a class is still super
useful, and I'll need the functionality to avoid changing even more of
Macaroni's own source code.

Fixing it turned out to be simple once I looked at the parser with a
clear head; basically there was some code that prevented a function
from being given to a target if a class was being parsed, but this didn't
check to make sure the function wasn't ~global.

Also changed back some tests I'd altered to get past this previously.