Create implicit ~units in parser; fix global fn issues
This is a huge rambling commit that got started as I went to fix some
issues where global functions where being assigned their own unit
targets by the Porg generator, the Lua plugin that adds unit targets to
nodes in a library that lack them..

I fixed this by making the PippyParser create implicit units each time
it sees a class or typedef (actually, to work around the ~hfile keyword
it has to create one a short time after- which really is obnoxious but
still works. I'll kill ~hfile one of these days).

Because of this, ~global functions are automatically added to the correct
unit (they just use whatever unit they're in).

This change also adds logic so a target knows if it was created
implicitly or not, which is useful for error messages. For example if a
class is created with an implicit unit, then later is partially defined
again under an explicit ~unit, the parser complains. The reverse also
holds true (as usual there are zero tests for this because this project
is spiraling out of control, but once M4 compiles in M4 it should be
possible to add C++ unit tests... finally).

Also got into the weeds fixing the stupid SimpleStringException class to
just propagate an entire string, and show that string even when "what()" was
called (before it only did this for the special method "message()", which
the LuaGlue was not calling, leading to crappy error messages that were
driving me mad).

There was also a notable bug fix where Scope::SwitchOwner was not
removing an Element from the list of Element's associated with the
Target it had originally been attached to. The effect was that the Porg
organizer (which still needs to run to generate units for the classes
the Lua plugin creates outside of the parser) would iterate elements
owned by a library and see all of these Element's that had been given
their own targets already. In the process it would duplicate units. If
the name was the same, there'd be an error message that the target
existed, but otherwise everything would work until linker time at which
point the Boost Build script would build two objects to the same source
file (one for each ~unit it had looked at). The fix was ridiculously
simple... too bad about those unit tests...